Create leapyear.java
Event Scheduling for Leap Year Celebrations
In a city, special events are scheduled every leap year to celebrate community milestones. Event organizers need a program that can help them identify whether a given year is a leap year, so they can schedule their events accordingly. If the year is a leap year, the program should notify the organizers to "Schedule Event"; otherwise, it should inform them "No Event This Year."



Problem Statement

Write a program to determine if a given year is a leap year. If it is, print "Schedule Event"; otherwise, print "No Event This Year."

A year is considered a leap year if:

It is divisible by 4,
If it is a century year (i.e., divisible by 100), it must also be divisible by 400.


Input Format

A single integer representing the year.



Output Format

Output "Schedule Event" if the year is a leap year.

Output "No Event This Year" if it is not a leap year.