Create apgp1.java
Oranges from the Growing Tree
A teacher tells a story about an orange tree. The tree gives 6 oranges for the first time, 11 oranges the second time, and then the tree decides to give double the difference between the previous two quantities (6 and 11), in addition to 11 oranges, and so on.

Your task is to generate the first 'n' terms of the number of oranges the tree gives, based on the following pattern.



Input Format

An integer 'n' which denotes the number of terms to be printed in the series.


Output Format

Print the series, where each term is separated by a space.