Create starpattern1.java
Building a Digital Staircase for a Smart Home System
Imagine you are working on a smart home project where a staircase's LED lighting system needs to be programmed to illuminate in a "right half pyramid" pattern. Each step lights up sequentially, forming a staircase-like appearance when viewed from the side. Your task is to write a program that simulates this lighting effect by printing a right half pyramid pattern.

You are given the number of steps (or rows) of the staircase. Write a program to display the "right half pyramid" pattern, where each row contains a number of stars corresponding to the row number.



Input Format

The input consists of a single integer n that denotes the number of rows in the pyramid.


Output Format

The output should display the right half pyramid pattern using stars (*), where each row has a number of stars equal to the row number.